requirements.txt
aioboto3==8.3.0
aiobotocore==1.2.2
aiocontextvars==0.2.2
aiohttp==3.7.4.post0
aiohttp-cors==0.7.0
aioitertools==0.7.1
aredis==1.1.8
asgiref==3.6.0
async-generator==1.10
async-timeout==3.0.1
attrs==22.2.0
boto3==1.16.52
botocore==1.19.52
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.0.1
click==8.1.3
cryptography==37.0.4
exceptiongroup==1.1.0
fastapi==0.65.1
h11==0.14.0
idna==3.4
jmespath==0.10.0
multidict==6.0.4
outcome==1.2.0
pycparser==2.21
pydantic==1.10.4
pymongo==3.11.4
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.17.1
requests==2.28.2
s3transfer==0.3.7
selenium==4.3.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.14.2
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.14
urllib3-secure-extra==0.1.0
uvicorn==0.14.0
uvloop==0.15.2
wrapt==1.14.1
wsproto==1.2.0
yarl==1.8.2
slowapi==0.1.7